requirement.txt
keystoneauth1 ~= 4.4.0
python-keystoneclient ~= 4.4.0
python-barbicanclient  ~= 5.2.0